# Conflicts:
#	source/Talpa/Talpa_10_WebApp/Areas/Manager/Controllers/OutingController.cs
#	source/Talpa/Talpa_10_WebApp/Areas/Manager/Views/Outing/Create.cshtml
#	source/Talpa/Talpa_10_WebApp/Controllers/SuggestionController.cs
#	source/Talpa/Talpa_10_WebApp/RequestModels/SuggestionRequest.cs
#	source/Talpa/Talpa_10_WebApp/ViewModels/OutingViewModel.cs
#	source/Talpa/Talpa_10_WebApp/Views/Suggestion/Create.cshtml
#	source/Talpa/Talpa_10_WebApp/Views/Suggestion/Edit.cshtml
#	source/Talpa/Talpa_20_BusinessLogicLayer/Models/Suggestion.cs
#	source/Talpa/Talpa_30_DataAccessLayer/Migrations/DataContextModelSnapshot.cs

+++ source/Talpa/Talpa_20_BusinessLogicLayer/Services/SuggestionService.cs
@@ -18,19 +18,37 @@ public class SuggestionService : ISuggestionService
         return _suggestionRepository.Create(suggestion, userid);
     }
 
-    public bool Update(Suggestion suggestion)
+    public bool Update(Suggestion newSuggestion, string userId)
     {
-        return _suggestionRepository.Update(suggestion);
+        Suggestion? suggestion = _suggestionRepository.GetById(newSuggestion.Id);
+        if (suggestion?.UserId == userId)
+        {
+            return _suggestionRepository.Update(newSuggestion);
+        }
+
+        return false;
     }
 
-    public bool Delete(int id)
+    public bool Delete(int id, string userId)
     {
-        return _suggestionRepository.Delete(id);
+        Suggestion? suggestion = _suggestionRepository.GetById(id);
+        if (suggestion?.UserId == userId)
+        {
+            return _suggestionRepository.Delete(id);
+        }
+
+        return false;
     }
 
-    public Suggestion? GetById(int id)
+    public Suggestion? GetById(int id, string userid)
     {
-        return _suggestionRepository.GetById(id);
+        Suggestion? suggestion = _suggestionRepository.GetById(id);
+        if (suggestion?.UserId == userid)
+        {
+            return suggestion;
+        }
+
+        return null;
     }
 
     public List<Suggestion> GetByIds(List<int> ids)
@@ -46,5 +64,10 @@ public class SuggestionService : ISuggestionService
     public List<Suggestion> GetAllByUserId(string id)
     {
         return _suggestionRepository.GetAllByUserId(id);
+    }
+
+    public bool Exists (string name)
+    {
+        return _suggestionRepository.Exists(name);
     }
 }
